Changed initial type to int

+++ model/memory.py
@@ -3,7 +3,7 @@ import torch.nn.functional as F
 
 class MemoryBank:
     def __init__(self, size, feature_dim, device):
-        self.bank = torch.zeros(size, feature_dim, device=device)
+        self.bank = torch.zeros(size, feature_dim, dtype=torch.int, device=device)
         self.labels = torch.empty(size, dtype=torch.long, device=device)
         self.ptr = 0
